Converetd la/lng coordinates from Long to Double References #148

+++ grails-app/domain/com/sra/searchfda/domain/Query.groovy
@@ -8,8 +8,8 @@ import groovy.transform.ToString
 class Query {
 
 	String search
-	Long lat
-	Long lng
+	Double lat
+	Double lng
 
 	static final int SEARCH_MAX_SIZE = 255
